+++ frontend/sass/components/_index.scss
@@ -1,4 +1,4 @@
 @forward "./buttons";
 @forward "./form";
 @forward "./section";
-@forward "./dropdownMenu";
+@forward "./dropdown";
